requirements.txt
flask
pyjwt
datetime
uuid
Flask-SQLAlchemy
